make web3container enforce connection before rendering childen
